Correct bug in minimum score computation
This commit corrects bug in 'chain_min_score'. The previous assumption was
that extra errors could be counted from long gaps between seeds. The
assumption was wrong since a big gap can be masked by a long MEM and it
does not guarantee that there is more than one mismatch in that
region. The new behaviour only adds one mismatch per gap or break in the
seed chain.