Fix warning that parameter u is not in range (0,1)
The error occurred when the reference contained "N", which forced the
the function 'quality()' to bail out and return a default value equal
to 0. The error is now caught by the caller and u is replaced by
a "worst case" value, with the intent to not get an over-optimistic
estimate of the mapping quality.

This commit fixes #2.