Improve quality estimate in the "good" case
The "good" case is when there is evidence that the target has no
duplicate. When that happens, the mapping quality requires a correction
to be well calibrated. The current method is based on the uncertainty
of estimating u and N0, and on the application of Bayes formula. This
commit also fixes a defect in the output when the input is given in
fasta format.