Implement dual workflow using MEM and skip seeds
Skip seeds are usually more sensitive than MEM seeds, so we can improve
the mapping quality by switching method. First tests showed that it is
more efficient to map first (using MEM seeds), estimate the quality and
then map again (using skip seeds) when the quality is too low.

The tests revealed that estimating the quality is approximately as long
as mapping. Most of the time is spent in the backward search and the
estimation accounts for up to one third of the calls to `get_rank()`.
In this commit, the process was optimized somewhat using a shortcut to
finish the estimation without all the calls to `get_rank()` as soon as
it is clear that a single hit remains in the bacwkard search. This
shaves off about 25% of the time spent in computing the mapping quality
(the total running time is around 15% faster).

This commit fixes #5.