Correct bug in skip seeds and improve chaining
This commit corrects a bug in skip seeding and several bugs in skip seed
chaining. Skip seeds are now designed to span the full read, so if the
last seed leaves a gap at the end of the read, an additional seed is
added, even if it does not respect the skip offset. The operation to
compute the minimum score of a skip seed chain has been
corrected. Chaining is now aborted (no alignments are performed) if a
single skip seed has more than MAX_SKIP_CHAIN_SEEDS loci.