Compress genome at indexing stage only
This commit updates the index structure. The genome nucleotides are now
compressed only once during indexing and stored in a .dna file. When the
genome is loaded, the .dna file is mmapped and its pointer is kept in
the index_t structure. To recover the original DNA sequence, a new
function decompress_genome() has been implemented. In practice, such
decompression is only used right before aligning the query
sequence. Previously, the genomic sequence of the best alignments were
stored (as pointers to the complete genome array) in the alignment
stack. Since the complete genome array is not available anymore, these
sequences are decompressed and allocated in an independent array at the
end of the alignment stage.