Deinterleave OCC structure
This commit updates the structure of the index. The occ table has been
updated to store the counts of each nucleotide in a big contiguous
block, instead of having interleaved bit fields and absolute count
markers. It also updates the default OCC interval size to 6, so that now
we expect an average of 2 'popcountl' operations per 'get_rank()'
query.

Additionally, all the remaining non-freed memory is explicitly freed
before the execution terminates.