Align MEM seeds without chaining
This commit updates the MEM strategy. MEM seeds are no longer chained
before alignment. Since no assumptions can be made on the effect of long
gaps in MEM chains, removing seeds from a MEM chain usually results in a
decrease of the chain minimum score. For this reason, it makes no sense
to spend computational resources computing SA and chaining MEMs whose
minscore will be anyway beaten by a single short MEM at the beginning of
the read (which have minscore 1). This update may have a negative effect
on the performance of the algorithm, but the MEM search is now
exhaustive. Previous versions of the MEM algorithm were skipping required
alignments derived from individual components of the MEM chains.

The new algorithm works as follows:

1. Separate MEMs by alignable and non-alignable.
2. Compute the minimum score of non-alignable MEMs.
3. Sort alignable MEMs by MEM span in the read.
4. Loop: Align alignable MEMs in all their loci (checking minscore).
5. Finally, return best alignment if better than non-alignable minscore.

* Non-alignable MEMs are those with more than MAX_MEM_SEED_LOCI
references to the Suffix Array.
* MEMs are aligned only if their minscore is smaller than the current
best score or if their minscore is equal to the current best score and
MAX_MINSCORE_REPEATS has not yet been reached.
* No assumptions can be made on non-alignable MEMs, therefore the best
score is only considered if it is strictly smaller than non-alignable
minscore.