Fix skip seeds minimum alignment score
This commit fixes a bug in the computation of minscore for skip seeds.
Previously, minscore was computed as the division between the gap length
in a chain of skip seeds and the distance between a skip seed and the
next non-overlapping skip seed. This was incorrect since the seeds are
not necessarily equally spaced. The minimum score is now computed on the
fly considering how many mismatches are needed to remove all the seeds
that form the gap. The code has not yet been thoroughly tested, be aware
that additional bugs may be found in this computation.