Fix bug in skip seed chaining
This commit fixes a bug in skip seed chaining algorithm. Previously,
seeds were chained based only on their respective distances in the read
and the reference. However, the algorithm was not checking that seeds
were chained following a strict order in their read position. This was
problematic in short repeats, where the same seed was pointing many
times in the same localized region of the reference. The current
behavior ensures that reads can only be chained if thay are also
consecutive in the read, not ony in the reference.