Fix bug in skip seeding
This commit fixes a bug in skip seeding. The previous version aborted
the seeding of the last (fill) seed if the previous one failed. This
potentially caused a loss in accuracy and errors in the computation of
the minimum score during seed chaining. This bug was likely caused
because the code for MEM seeding was reused in skip seeds.