Fix bug in ranged CSA query
This commit fixes a bug in 'query_csa_range()'. The function starts a
recursive call to 'recursive_csa_query()', which assumes that the loci
range has length strictly greater than one. When the function is called
with a seed with only one target, it fails to find the proper reference
position. The fix redirects the call to 'query_csa()' if the range to
be queried has length 1.