Fix bug in 'query_csa_range()'
This commit finally fixes the bug referenced by the previous commit.
The function 'query_csa_range()' could actually accept ranges of
length 1 but it did not check whether the 'bwt->zero' position was
within the current inspected range. If 'bwt->zero' is in the queried
range any attempt to backward search passed this point would fail.
The bug was therefore affecting the sequences between the beginning
of the reference and the first CSA-marker, unless 'bwt->zero' was a
CSA-marked position.