Implement chunk-merge SA construction
This commit implements a new strategy for index construction. SA is now
computed in chunks to minimize the memory usage. At the chunk merging
step, all the other structures (occ, csa and bwt) are computed on the
fly. The structure of occ has been updates and now the different symbol
intervals are interleaved. This requires an update of the occ query
algorithms.